Moved prom language features to datasource language provider
