+++ public/app/plugins/datasource/prometheus/promql.ts
